Fixed latest correction issues

+++ srcs/main_bonus.c
@@ -6,7 +6,7 @@
 /*   By: dkeraudr <dkeraudr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/16 11:58:38 by dkeraudr          #+#    #+#             */
-/*   Updated: 2023/07/10 19:01:24 by dkeraudr         ###   ########.fr       */
+/*   Updated: 2023/07/11 19:25:08 by dkeraudr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	main(int argc, char **argv, char **envp)
 	t_pipex	*pipex;
 	int		status;
 
-	if (argc < 6)
+	if (argc < 5)
 	{
 		ft_error("usage: ./pipex_bonus file1 cmd1 cmd2 file2");
 		ft_error("or");
@@ -32,7 +32,8 @@ int	main(int argc, char **argv, char **envp)
 	status = ft_pipex(pipex);
 	if (status == -1)
 		return (ft_error("pipex failed"));
-	unlink(pipex->file_in);
+	if (pipex->here_doc)
+		unlink(pipex->file_in);
 	ft_free_pipex(pipex);
 	return (status);
 }
